requirements.txt
Django
django-extensions
django-model-utils
django-debug-toolbar
ipython
